requirements.txt
Flask==0.12.2
Flask-DbShell==1.0
Flask-GCM==0.2.0
Flask-Migrate==2.0.3
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
mysql-connector-python-rf==2.2.2
nose==1.3.7
Pillow==4.1.1
python-dateutil==2.6.0
Sphinx==1.6.1
Sphinxcontrib-httpdomain==1.5.0
ujson==1.35
